+++ android/src/main/java/org/qooapps/wear_os_plugin/WearOsPlugin.java
@@ -1,17 +1,28 @@
 package org.qooapps.wear_os_plugin;
 
 import android.app.Activity;
+import android.app.Application;
 import android.content.Context;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.os.Build;
+import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
 import android.os.VibrationEffect;
 import android.os.Vibrator;
+import android.util.Log;
+import android.view.ActionMode;
+import android.view.KeyEvent;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.MotionEvent;
+import android.view.SearchEvent;
 import android.view.View;
+import android.view.Window;
+import android.view.WindowManager;
+import android.view.accessibility.AccessibilityEvent;
 
 import androidx.annotation.NonNull;
 
@@ -33,7 +44,9 @@ import io.flutter.plugin.common.MethodChannel.Result;
  */
 public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHandler {
     static final String CHANNEL_METHOD = "wear_os_plugin/method";
-    static final String CHANNEL_EVENT = "wear_os_plugin/event";
+    static final String CHANNEL_MOTION_EVENTS = "wear_os_plugin/motionEvents";
+    static final String CHANNEL_KEY_EVENTS = "wear_os_plugin/keyEvents";
+    static final String CHANNEL_LIFECYCLE_EVENTS = "wear_os_plugin/lifecycleEvents";
 
     /// The MethodChannel that will the communication between Flutter and native Android
     ///
@@ -41,8 +54,10 @@ public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHan
     /// when the Flutter Engine is detached from the Activity
     private Context mContext;
     private MethodChannel channel;
-    private static EventChannel.EventSink mEventSink;
-    private static final Handler mHandler = new Handler();
+    private EventChannel.EventSink mMotionEventsSink;
+    private EventChannel.EventSink mKeyEventsSink;
+    private EventChannel.EventSink mLifecycleEventsSink;
+    private final Handler mHandler = new Handler();
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
@@ -50,15 +65,39 @@ public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHan
         channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), CHANNEL_METHOD);
         channel.setMethodCallHandler(this);
 
-        new EventChannel(flutterPluginBinding.getBinaryMessenger(), CHANNEL_EVENT).setStreamHandler(new EventChannel.StreamHandler() {
+        new EventChannel(flutterPluginBinding.getBinaryMessenger(), CHANNEL_MOTION_EVENTS).setStreamHandler(new EventChannel.StreamHandler() {
             @Override
             public void onListen(Object o, EventChannel.EventSink eventSink) {
-                mEventSink = eventSink;
+                mMotionEventsSink = eventSink;
             }
 
             @Override
             public void onCancel(Object o) {
-                mEventSink = null;
+                mMotionEventsSink = null;
+            }
+        });
+
+        new EventChannel(flutterPluginBinding.getBinaryMessenger(), CHANNEL_KEY_EVENTS).setStreamHandler(new EventChannel.StreamHandler() {
+            @Override
+            public void onListen(Object o, EventChannel.EventSink eventSink) {
+                mKeyEventsSink = eventSink;
+            }
+
+            @Override
+            public void onCancel(Object o) {
+                mKeyEventsSink = null;
+            }
+        });
+
+        new EventChannel(flutterPluginBinding.getBinaryMessenger(), CHANNEL_LIFECYCLE_EVENTS).setStreamHandler(new EventChannel.StreamHandler() {
+            @Override
+            public void onListen(Object o, EventChannel.EventSink eventSink) {
+                mLifecycleEventsSink = eventSink;
+            }
+
+            @Override
+            public void onCancel(Object o) {
+                mLifecycleEventsSink = null;
             }
         });
     }
@@ -72,20 +111,39 @@ public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHan
     @Override
     public void onAttachedToActivity(ActivityPluginBinding binding) {
         Activity activity = binding.getActivity();
+        // get all lifecycle events:
+        activity.registerActivityLifecycleCallbacks(new LifecycleCallbacks(this));
         // using transparent background so the round screen will be seen:
         activity.getIntent().putExtra("background_mode", "transparent");
         // set motion (means rotary) event listener for main view:
-        if (activity.getWindow() != null) {
-            View mainView = activity.getWindow().findViewById(android.R.id.content);
-            if (mainView == null) mainView = activity.getWindow().getDecorView();
+        Window window = activity.getWindow();
+        if (window != null) {
+            /*
+            View mainView = window.findViewById(android.R.id.content);
+            if (mainView == null) mainView = window.getDecorView();
             if (mainView.getRootView() != null) mainView = mainView.getRootView();
+            final View view = mainView;
 
+            final WearOsPlugin plugin = this;
             if (mainView != null) {
                 // mainView.setAlpha(0.5f);
-                mainView.setOnGenericMotionListener((view, motionEvent) -> {
-                    return WearOsPlugin.onGenericMotionEvent(motionEvent);
+                mainView.setOnGenericMotionListener((v, motionEvent) -> {
+                    Log.i("EVENT", "motionEvent");
+                    plugin.sendGenericMotionEvent(motionEvent);
+                    return true;
+                });
+                mainView.setOnKeyListener(new View.OnKeyListener() {
+                    @Override
+                    public boolean onKey(View view, int i, KeyEvent keyEvent) {
+                        Log.i("EVENT", "keyEvent");
+                        return false;
+                    }
                 });
             }
+            */
+
+            // get all motion and key events:
+            window.setCallback(new WindowCallbacks(window.getCallback(), this));
         }
     }
 
@@ -125,53 +183,33 @@ public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHan
             }
             break;
             case "isScreenRound": {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                    Configuration c = mContext.getResources().getConfiguration();
-                    result.success(c.isScreenRound());
-                } else {
-                    result.success(null);
-                }
+                Configuration c = mContext.getResources().getConfiguration();
+                result.success(c.isScreenRound());
             }
             break;
             case "vibrate": {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                    int duration = call.argument("duration");
-                    int amplitude = call.argument("amplitude");
-                    String effect = call.argument("effect");
-                    Vibrator v = (Vibrator) mContext.getSystemService(Context.VIBRATOR_SERVICE);
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                        switch (Objects.requireNonNull(effect)) {
-                            case "click": {
-                                v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_CLICK));
-                            }
-                            break;
-                            case "tick":
-                                v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_TICK));
-                                break;
-                            case "double_click":
-                                v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_DOUBLE_CLICK));
-                                break;
-                            case "heavy_click":
-                                v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_HEAVY_CLICK));
-                                break;
-                            default:
-                                v.vibrate(VibrationEffect.createOneShot(duration, amplitude));
-                                break;
+                int duration = call.argument("duration");
+                int amplitude = call.argument("amplitude");
+                String effect = call.argument("effect");
+                Vibrator v = (Vibrator) mContext.getSystemService(Context.VIBRATOR_SERVICE);
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                    switch (Objects.requireNonNull(effect)) {
+                        case "click": {
+                            v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_CLICK));
                         }
-                    } else {
-                        switch (Objects.requireNonNull(effect)) {
-                            case "double_click":
-                            case "heavy_click":
-                            case "tick":
-                            case "click": {
-                                v.vibrate(VibrationEffect.createOneShot(20, 50));
-                            }
+                        break;
+                        case "tick":
+                            v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_TICK));
                             break;
-                            default:
-                                v.vibrate(VibrationEffect.createOneShot(duration, amplitude));
-                                break;
-                        }
-
+                        case "double_click":
+                            v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_DOUBLE_CLICK));
+                            break;
+                        case "heavy_click":
+                            v.vibrate(VibrationEffect.createPredefined(VibrationEffect.EFFECT_HEAVY_CLICK));
+                            break;
+                        default:
+                            v.vibrate(VibrationEffect.createOneShot(duration, amplitude));
+                            break;
                     }
                 }
                 result.success(null);
@@ -182,19 +220,20 @@ public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHan
                 break;
         }
     }
-/*
-    private void sendSuccessMsgToEventChannel(Object msg) {
-        if (mEventSink != null)
-            runOnMainThread(() -> mEventSink.success(msg));
-    }
 
-    private void sendFailMsgToEventChannel(String errCode, String errMsg, Object errDetail) {
-        if (mEventSink != null)
-            runOnMainThread(() -> mEventSink.error(errCode, errMsg, errDetail));
-    }
-*/
+    /*
+        private void sendSuccessMsgToEventChannel(Object msg) {
+            if (mEventSink != null)
+                runOnMainThread(() -> mEventSink.success(msg));
+        }
+
+        private void sendFailMsgToEventChannel(String errCode, String errMsg, Object errDetail) {
+            if (mEventSink != null)
+                runOnMainThread(() -> mEventSink.error(errCode, errMsg, errDetail));
+        }
+    */
     // start runnable in main thread:
-    private static void runOnMainThread(Runnable runnable) {
+    private void runOnMainThread(Runnable runnable) {
         if (Looper.myLooper() == Looper.getMainLooper()) {
             runnable.run();
         } else {
@@ -202,15 +241,28 @@ public class WearOsPlugin implements FlutterPlugin, ActivityAware, MethodCallHan
         }
     }
 
-    public static boolean onGenericMotionEvent(MotionEvent event) {
-        if (mEventSink != null) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                Map<String, Object> msg = new HashMap<>(1);
-                msg.put("scroll", event.getAxisValue(MotionEvent.AXIS_SCROLL));
-                runOnMainThread(() -> mEventSink.success(msg));
-                return true;
-            }
+    public void sendGenericMotionEvent(MotionEvent event) {
+        if (mMotionEventsSink != null) {
+            Map<String, Object> msg = new HashMap<>(1);
+            msg.put("scroll", event.getAxisValue(MotionEvent.AXIS_SCROLL));
+            runOnMainThread(() -> mMotionEventsSink.success(msg));
         }
-        return false;
+    }
+
+    public void sendKeyEvent(KeyEvent event) {
+        if (mKeyEventsSink != null) {
+            Map<String, Object> msg = new HashMap<>(2);
+            msg.put("keyCode", event.getKeyCode());
+            msg.put("down", event.getAction() == KeyEvent.ACTION_DOWN);
+            runOnMainThread(() -> mKeyEventsSink.success(msg));
+        }
+    }
+
+    public void sendLifecycleEvent(String action) {
+        if (mLifecycleEventsSink != null) {
+            Map<String, Object> msg = new HashMap<>(1);
+            msg.put("action", action);
+            runOnMainThread(() -> mLifecycleEventsSink.success(msg));
+        }
     }
 }